+++ NEWS.md
@@ -1,10 +1,21 @@
+# webr version 0.1.1
+=====================
+(19-May-2018)
+
+* new functions freqTable() and myFreqTable() added
+
+* dictionary file dicTable added
+
+* new function langchoice1() added
+
+
 # webr version 0.1.0
 =====================
 (18-Apr-2018)
 
 * new function plot.htest() added
 
-* new function rens.text() and cox.stuart.test() added
+* new functions rens.text() and cox.stuart.test() added